rounding for financing rate, profile h2mem, junit

+++ src/main/java/lu/crx/financing/model/entities/FactoredInvoice.java
@@ -28,7 +28,7 @@ public class FactoredInvoice {
     @Column(nullable = true)
     private LocalDate fundingDate;
     private long financingTerm;
-    private double financingRate;
+    private long financingRate;
     private long earlyPaymentAmount;
